+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 ^^^^^^^^^^^^^^^^^^
 
 - Allow custom serialization in JSONRenderer without abusing __repr__.
+- Enhance Twisted support by offering JSONification of non-structlog log entries.
 
 
 0.1.0 (2013-09-12)